+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## 2020.2.6
 
+#### Added
+
+* Handle Drop in Callable Reorderable Lists : Dropped Callables components are added to the list. Dropped Game Objects will prompt a menu to select which callable to add.
+
+#### Fixed
+
+* Fixed Callable Reorderable List Undo.
+
 ## 2020.2.5
 
 #### Added